Move call wrappers to separate file.
